+++ tests/MaxMind/Test/MinFraud/Model/DeviceTest.php
@@ -15,6 +15,7 @@ class DeviceTest extends \PHPUnit_Framework_TestCase
             'confidence' => 99,
             'id' => '915d5202-d6c5-4616-b2c1-87683975dadb',
             'last_seen' => '2016-06-08T14:16:38Z',
+            'local_time' => '2016-06-10T14:19:10-08:00',
         ];
         $device = new Device($array);
 
@@ -34,6 +35,12 @@ class DeviceTest extends \PHPUnit_Framework_TestCase
             $array['last_seen'],
             $device->lastSeen,
             'last_seen'
+        );
+
+        $this->assertSame(
+            $array['local_time'],
+            $device->localTime,
+            'local_time'
         );
 
         $this->assertSame(